# backend/ApiSpec.yaml
openapi: 3.0.3
info:
  title: Voting DApp Backend API
  description: Express/TypeScript backend enhancement for gconnect/voting-dapp-pyteal-react
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development
paths:
...
  /api/voting/results/{appId}:
    get:
      summary: Get results for a specific application ID
      parameters:
        - name: appId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Election results
...
  /api/voting/verify/{txId}:
    get:
      summary: Verify a vote transaction on the blockchain
      parameters:
        - name: txId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction verified
        '404':
          description: Transaction not found
...